requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
GitPython==3.1.40
PyYAML==6.0.1
APScheduler==3.10.4
requests==2.31.0
python-engineio==4.9.1
python-socketio==5.11.1
eventlet==0.35.2
gevent==23.9.1
gevent-websocket==0.10.1
dnspython==2.4.2 